requirements.txt
Flask==2.3.2
pandas==2.1.2
PyPDF2==3.0.1
python-docx==0.8.11
gunicorn==21.2.0
